// src/nova/generated/pch-profiles.ts
// Generated by tools/compiler/build-pch.mjs; regenerate after changing the engine or profiles.
export const pchProfiles = [
  {
    "headers": [
      "iostream"
    ],
    "prelude": "#include <iostream>\n",
    "filename": "iostream-5b8b3734b47d32bd.pch.bin",
    "sha256": "5b8b3734b47d32bd3182404707f3786a1376800b099aef4efa27d05701b8b649",
    "bytes": 12686096,
    "compressedBytes": 7287957
  },
  {
    "headers": [
      "iostream",
      "vector"
    ],
    "prelude": "#include <iostream>\n#include <vector>\n",
    "filename": "iostream-vector-893905b26e360426.pch.bin",
    "sha256": "893905b26e3604265486b6f225c937e33d3e8a4361de8eb02c569eb67c46fd00",
    "bytes": 13519304,
    "compressedBytes": 7826845
  }
] as const